refactor(reader): divide controls into files
